Standardize call to get_shipping_methods
get_shipping_methods was called with different parameters in different
parts of the code. This commit now changes things to call it with as
much information as is available at the time, hence always including the
request and the basket. It also uses named keyword arguments to make
backwards compatibility easier (as the method signature obviously
changed).

Closes #99.