Add sandbox' shipping migrations
We discovered that the sandbox would break on PostgreSQL, because we
forgot to follow the Oscar instructions to fork an app, which includes
copying migrations across.

In an ideal world, we would somehow test that this doesn't happen again.
But as this only affects developers of django-oscar-paypal, and not
general users, and this is the first issue like this,
I'm not bothering at this point.