Preserve phone numbers across PayPal Express transactions
When stores collect customers' phone numbers, they may do so for good
reason and may want them preserved on order confirmations and
receipts. The PayPal Express views previously would not forward this
information to PayPal (despite an optional field,
PAYMENTREQUEST_n_SHIPTOPHONENUM, being available for this purpose in
the NVP API). Since in the confirmation view the shipping address is
constructed from the PayPal response, not the previously entered user
data, this would mean that the phone number would silently get lost in
the transaction.

Fix PayPal Express so that user address / shipping address fields are
correctly translated into the PAYMENTREQUEST_n_SHIPTOPHONENUM param on
send, and then re-initialize the shipping address using this
information on receive.